+++ src/renderer/widgets/RemoveVotesModal/index.ts
@@ -1,2 +1,3 @@
 export { RemoveVotesModal } from './ui/RemoveVotesModal';
+export { RemoveVotesShardsModal } from './shards/components/RemoveVotesShardsModal';
 export { removeVotesModel } from './model/removeVotesModal';
